backend/requirements.txt
# 핵심 웹 프레임워크
fastapi==0.104.1
uvicorn==0.24.0

# 데이터 검증 및 설정
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.0
email-validator==2.1.0

# 보안 및 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP 클라이언트 및 파일 처리
httpx==0.25.2
aiofiles==23.2.1

# 데이터베이스 및 캐시
redis==5.0.1
azure-cosmos==4.5.1

# Azure 서비스 (선택사항 - 실제 사용 시에만 설치)
# azure-cognitiveservices-speech==1.34.0
# openai==1.3.8

# 이미지 처리 (Windows에서 문제 발생 시 주석 처리)
# pillow==10.1.0
